# lib/api-zod/src/generated/types/alertSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';

export interface AlertSettings {
  browserNotificationsEnabled: boolean;
  inAppNotificationsEnabled: boolean;
  minimumSeverity: AlertSeverity;
  /** @nullable */
  quietHoursStart: string | null;
  /** @nullable */
  quietHoursEnd: string | null;
  /** @nullable */
  timezone: string | null;
  globalOptOut: boolean;
}