# lib/api-zod/src/generated/types/alertSeverity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  info: 'info',
  watch: 'watch',
  alert: 'alert',
  critical: 'critical',
} as const;