# lib/api-zod/src/generated/types/pushSubscriptionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PushSubscriptionRequestKeys } from './pushSubscriptionRequestKeys';

export interface PushSubscriptionRequest {
  endpoint: string;
  keys: PushSubscriptionRequestKeys;
  /** @maxLength 120 */
  deviceLabel?: string;
}