# lib/api-zod/src/generated/types/updateAlertSettingsRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAlertSettingsRequestMinimumTier } from './updateAlertSettingsRequestMinimumTier';

export interface UpdateAlertSettingsRequest {
  browserNotificationsEnabled: boolean;
  minimumTier: UpdateAlertSettingsRequestMinimumTier;
  notifyOnWatch: boolean;
}