# lib/api-zod/src/generated/types/verifiedAlphaAlert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';

export interface VerifiedAlphaAlert {
  id: string;
  eventKey: string;
  symbol: string;
  severity: AlertSeverity;
  triggerReason: string;
  detectionState: string;
  confirmationStatus: string;
  alphaScore: number;
  /** @nullable */
  alphaVelocity30s?: number | null;
  confidence: number;
  /** @nullable */
  triggerPrice?: number | null;
  preBreakoutState: string;
  satisfiedEvidence: string[];
  missingEvidence: string[];
  generatedAt: Date;
  /** @nullable */
  readAt: Date | null;
  /** @nullable */
  acknowledgedAt: Date | null;
}